Comentários dos patterns dependencyInjection e iterator
