--- src/com/examples/dependencyInjection/IInjetor.java
@@ -1,10 +1,0 @@
-package com.examples.dependencyInjection;
-
-/**
- * Created by ricardo on 2/13/17.
- */
-public interface IInjetor {
-
-    public IServicoMsg getMessageService();
-
-}
