--- src/com/examples/dependencyInjection/IServicoMsg.java
@@ -1,5 +1,0 @@
-package com.examples.dependencyInjection;
-
-public interface IServicoMsg {
-    void sendMessage(String assunto, String mensagem);
-}
