+++ src/com/examples/iterator/Iterator.java
@@ -4,12 +4,13 @@ import java.util.Enumeration;
 import java.util.Hashtable;
 
 /**
- * Created by ricardo on 2/20/17.
+ * Essa classe implementa um iterador de objetos do tipo "Valor".
  */
 public class Iterator {
+    /*Os itens do tipo valor são armazenados em um objeto do tipo Hashtable
+    * o objeto indice armazena o indice atual do iterador enquanto o valor numItens
+    * guarda o total de itens armazenados*/
     private Hashtable itens;
-    private Enumeration e;
-    private int itemAtual;
     public int indice;
     public int numItens;
 
@@ -19,16 +20,19 @@ public class Iterator {
         numItens = 0;
     }
 
+    //Adiciona um item no iterador e soma o total de itens
     public int add(Object item) {
         itens.put(numItens, item);
         numItens++;
         return numItens;
     }
 
+    //Busca o item de indice 0 do iterador
     public Object getFirst() {
         return itens.get(0);
     }
 
+    //Método que verifica se o iterador chegou ao final
     public boolean isDone() {
         if (indice == numItens)
             return true;
@@ -36,12 +40,15 @@ public class Iterator {
             return false;
     }
 
+    /*Retorna o item de um indice especificado. Há uma verificação para caso o índice passado seja maior
+      que o número de itens. */
     public Object get(int index) {
         if (index > numItens)
             return null;
         return itens.get(index);
     }
 
+    //Navega para o proximo item da lista do iterador
     public boolean next() {
         if (indice < numItens) {
             indice++;
@@ -50,6 +57,7 @@ public class Iterator {
             return false;
     }
 
+    //Navega para o item anterior da lista do iterador
     public boolean last() {
         if (indice > 0) {
             indice--;
@@ -60,6 +68,7 @@ public class Iterator {
         }
     }
 
+    //Pega o item atual na navegaçao do iterador
     public Object getCurrentItem() {
         return itens.get(indice);
     }
